Splitted StorageManagement file
