+++ src/mtool/ui/src/containers/StorageManagement/index.js
@@ -31,54 +31,27 @@
  */
 
 import React, { Component, lazy, Suspense } from "react";
-import { Box, Grid, Typography, Paper, AppBar, Tabs, Tab, Select, FormControl, InputLabel, MenuItem, Tooltip, Button, Zoom } from "@material-ui/core";
+import { Box, Grid, Typography, AppBar, Tabs, Tab } from "@material-ui/core";
 import { withStyles, MuiThemeProvider as ThemeProvider } from "@material-ui/core/styles";
 import { connect } from "react-redux";
 import { Route, Switch, withRouter, Redirect } from 'react-router-dom';
-import io from "socket.io-client";
 import "react-dropdown/style.css";
 import "react-table/react-table.css";
-import InfoIcon from "@material-ui/icons/Info";
-import AutoCreate from "../../components/AutoCreate";
 import Header from "../../components/Header";
 import Sidebar from "../../components/Sidebar";
 import MToolLoader from "../../components/MToolLoader";
-import RebuildProgress from "../../components/RebuildProgress";
 import AlertDialog from "../../components/Dialog";
 import "./StorageManagement.css";
 import MToolTheme, { customTheme } from "../../theme";
 import SelectSubsystem from "../../components/SelectSubsystem";
-import Legend from "../../components/Legend";
 import * as actionTypes from "../../store/actions/actionTypes";
-import formatBytes from "../../utils/format-bytes";
 import getSubsystemForArray from "../../utils/subsystem";
-import LightTooltip from "../../components/LightTooltip";
-import { BYTE_FACTOR } from "../../utils/constants";
 
 
-const ArrayCreate = lazy(() => import("../../components/ArrayManagement/ArrayCreate"));
-const ArrayShow = lazy(() => import("../../components/ArrayManagement/ArrayShow"));
-const CreateVolume = lazy(() => import("../../components/VolumeManagement/CreateVolume"));
-const VolumeList = lazy(() => import("../../components/VolumeManagement/VolumeList"));
+const ArrayCreate = lazy(() => import("./ArrayCreate"));
+const ArrayManage = lazy(() => import("./ArrayManage"));
 
 const styles = (theme) => ({
-  dashboardContainer: {
-    display: "flex",
-  },
-  statsWrapper: {
-    display: "flex",
-    flexWrap: "wrap",
-    width: "100%",
-    boxSizing: "border-box",
-    zIndex: 100,
-    flexBasis: "100%",
-    height: "100%",
-    alignContent: "center",
-    padding: theme.spacing(0, 3),
-    [theme.breakpoints.down("sm")]: {
-      marginTop: theme.spacing(1),
-    },
-  },
   content: {
     flexGrow: 1,
     padding: theme.spacing(3),
@@ -88,137 +61,35 @@ const styles = (theme) => ({
     width: "calc(100% - 256px)",
     boxSizing: "border-box",
   },
-  statsContainer: {
-    margin: theme.spacing(1, 0, 2),
-  },
-  volumeStats: {
-    width: "100%",
-    border: "0px solid gray",
-    height: 50,
-  },
-  arraySelect: {
-    minWidth: 170,
-    "&>div>p": {
-      overflow: "hidden",
-      textOverflow: "ellipsis"
-    }
-  },
-  arraySelectStatus: {
-    padding: theme.spacing(0, 2),
-    [theme.breakpoints.down("sm")]: {
-      padding: theme.spacing(0, 1),
-    },
-    [theme.breakpoints.down("xs")]: {
-      justifyContent: "center",
-    }
-  },
-  arraySelectGrid: {
-    [theme.breakpoints.down("xs")]: {
-      textAlign: "center",
-      width: "100%"
-    }
-  },
-  arrayInfoContainer: {
-    position: "relative"
-  },
-  arrayInfoIcon: {
-    position: "absolute",
-    right: theme.spacing(2),
-    top: theme.spacing(2),
-    '&:hover': {
-      cursor: 'pointer'
-    }
-  },
-  selectForm: {
-    margin: theme.spacing(0.5, 2),
-    width: "60%",
-    minWidth: "170px",
-    [theme.breakpoints.down("xs")]: {
-      margin: theme.spacing(1, 0),
-      width: "80%",
-    },
-  },
   toolbar: customTheme.toolbar,
   titleContainer: {
     marginTop: theme.spacing(1),
   },
-  statusText: {
-    display: "flex",
-    height: "100%",
-    alignItems: "center",
-    margin: theme.spacing(0.5, 2),
-    width: "80%",
-    minWidth: "170px",
-    [theme.breakpoints.down("xs")]: {
-      margin: theme.spacing(1, 0),
-      display: "inline-flex",
-    },
-  },
-  volumeStatsPaper: {
-    height: "100%",
-    display: "flex",
-    position: "relative",
-    flexDirection: "column"
-  },
-  legendContainer: {
-    justifyContent: "flex-end"
-  },
-  selectedTab: customTheme.tab.selected,
   pageHeader: customTheme.page.title,
-  cardHeader: customTheme.card.header,
-  tooltip: {
-    backgroundColor: "#f5f5f9",
-    opacity: 1,
-    color: "rgba(0, 0, 0, 1)",
-    maxWidth: 220,
-    fontSize: theme.typography.pxToRem(12),
-    border: "1px solid #dadde9",
-    "& b": {
-      fontWeight: theme.typography.fontWeightMedium,
-    },
-  },
+  selectedTab: customTheme.tab.selected,
   card: {
     marginTop: theme.spacing(1),
   },
-  rebuildButton: {
-    cursor: "pointer",
-    marginLeft: theme.spacing(1)
-  }
 });
 
-// namespace to connect to the websocket for multi-volume creation
-const createVolSocketEndPoint = ":5000/create_vol";
 
 class StorageManagement extends Component {
   constructor(props) {
     super(props);
     this.state = {
       mobileOpen: false,
-      createVolSocket: io(createVolSocketEndPoint, {
-        transports: ["websocket"],
-        query: {
-          "x-access-token": localStorage.getItem("token"),
-        },
-      }),
       mountSubsystem: ""
     };
-    this.deleteVolumes = this.deleteVolumes.bind(this);
-    this.fetchVolumes = this.fetchVolumes.bind(this);
-    this.createArray = this.createArray.bind(this);
-    this.createVolume = this.createVolume.bind(this);
-    this.deleteArray = this.deleteArray.bind(this);
     this.alertConfirm = this.alertConfirm.bind(this);
     this.fetchDevices = this.fetchDevices.bind(this);
     this.handleTabChange = this.handleTabChange.bind(this);
     this.fetchMaxVolumeCount = this.fetchMaxVolumeCount.bind(this);
     this.handleDrawerToggle = this.handleDrawerToggle.bind(this);
     this.changeArray = this.changeArray.bind(this);
-    this.changeMountStatus = this.changeMountStatus.bind(this);
     this.mountConfirm = this.mountConfirm.bind(this);
     this.changeMountSubsystem = this.changeMountSubsystem.bind(this);
     this.closeMountPopup = this.closeMountPopup.bind(this);
     this.isSubsystemReserved = this.isSubsystemReserved.bind(this);
-    this.rebuildArray = this.rebuildArray.bind(this);
     const urlParams = new URLSearchParams(window.location.search);
     const array = urlParams.get("array");
     if (array) {
@@ -227,18 +98,9 @@ class StorageManagement extends Component {
   }
 
   componentDidMount() {
-    this.props.Get_Config();
     this.fetchDevices();
     this.fetchMaxVolumeCount();
     this.props.Get_Subsystems();
-  }
-
-  componentDidUpdate() {
-    if (window.location.href.indexOf('manage') > 0
-      && window.location.href.indexOf(`array=${this.props.selectedArray}`) < 0) {
-      this.props.history.push(`/storage/array/manage?array=${this.props.selectedArray}`);
-      this.fetchVolumes();
-    }
   }
 
   handleDrawerToggle() {
@@ -257,25 +119,6 @@ class StorageManagement extends Component {
     }
   }
 
-  createVolume(volume) {
-    this.props.Create_Volume({
-      name: volume.volume_name,
-      size: volume.volume_size,
-      description: volume.volume_description,
-      unit: volume.volume_units,
-      maxbw: volume.maxbw,
-      maxiops: volume.maxiops,
-      minbw: volume.minbw,
-      miniops: volume.miniops,
-      count: volume.volume_count,
-      subsystem: volume.subsystem,
-      suffix: volume.volume_suffix,
-      stop_on_error: volume.stop_on_error_checkbox,
-      mount_vol: volume.mount_vol,
-      max_available_size: this.props.arrayMap[this.props.selectedArray].totalsize - this.props.arrayMap[this.props.selectedArray].usedspace,
-    });
-  }
-
   changeArray(event) {
     const { value } = event.target;
     this.props.history.push(`/storage/array/manage?array=${value}`);
@@ -283,29 +126,17 @@ class StorageManagement extends Component {
     this.props.Get_Volumes({ array: value });
   }
 
-  fetchVolumes() {
-    this.props.Get_Volumes({ array: this.props.selectedArray });
-  }
 
   fetchMaxVolumeCount() {
     this.props.Get_Max_Volume_Count();
   }
 
-  deleteVolumes(volumes) {
-    const vols = [];
-    volumes.forEach((volume) => {
-      vols.push({ name: volume.name, isMounted: volume.status === "Mounted" });
-    });
-    this.props.Delete_Volumes({ volumes: vols });
-  }
 
   fetchDevices() {
     this.props.Get_Devices(this.props.history);
   }
 
-  createArray(array) {
-    this.props.Create_Array(array);
-  }
+
 
   closeMountPopup() {
     this.setState({
@@ -356,58 +187,16 @@ class StorageManagement extends Component {
     });
   }
 
-  changeMountStatus(payload) {
-    if (payload.status !== "Mounted") {
-      this.mountConfirm(payload);
-    } else {
-      this.props.Change_Mount_Status(payload);
-    }
-  }
 
-  deleteArray() {
-    this.props.Delete_Array({ arrayname: "" });
-  }
+
+
 
   alertConfirm() {
     this.props.Close_Alert();
   }
 
-  rebuildArray() {
-    this.props.Rebuild_Array(this.props.selectedArray);
-  }
 
   render() {
-    let totalVolSize = 0;
-    for (let i = 0; i < this.props.volumes.length; i += 1) {
-      totalVolSize += this.props.volumes[i].size;
-    }
-    const volumeFilledStyle = {
-      width: `${this.props.arrayMap[this.props.selectedArray] && this.props.arrayMap[this.props.selectedArray].totalsize !== 0
-        ? (100 * totalVolSize) / this.props.arrayMap[this.props.selectedArray].totalsize
-        : 0
-        }%`,
-      height: "100%",
-      backgroundColor: "rgba(51, 158, 255,0.6)",
-      float: "left",
-      overflowY: "hidden",
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "center",
-    };
-    const volumeFreeStyle = {
-      width: `${this.props.arrayMap[this.props.selectedArray] && this.props.arrayMap[this.props.selectedArray].totalsize !== 0
-        ? 100 - (100 * totalVolSize) / this.props.arrayMap[this.props.selectedArray].totalsize
-        : 100
-        }%`,
-      height: "100%",
-      color: "white",
-      backgroundColor: "rgba(0, 186, 0, 0.6)",
-      float: "left",
-      overflowY: "hidden",
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "center",
-    };
     const { classes } = this.props;
 
     return (
@@ -458,282 +247,25 @@ class StorageManagement extends Component {
                   <Redirect exact from="/storage/array/" to="/storage/array/create" />
                   <Route path="/storage/array/create">
                     <Grid container spacing={1} className={classes.card}>
-                      <Grid item xs={12}>
-                        <Paper spacing={3} className={classes.spaced}>
-                          <Grid container justifyContent="space-between">
-                            <ArrayCreate
-                              createArray={this.createArray}
-                              config={this.props.config}
-                              selectedRaid={this.props.selectedRaid}
-                              selectRaid={this.props.Select_Raid}
-                              disks={this.props.ssds}
-                              data-testid="arraycreate"
-                              metadisks={this.props.metadisks}
-                              diskDetails={this.props.diskDetails}
-                              getDiskDetails={this.props.Get_Disk_Details}
-                            />
-                          </Grid>
-                        </Paper>
-                        <AutoCreate
-                          disks={this.props.ssds}
-                          metadisks={this.props.metadisks}
-                          autoCreateArray={this.props.Auto_Create_Array}
-                          config={this.props.config}
-                        />
-                        {(this.props.posMountStatus === "EXIST_NORMAL") ? (
-                          <Typography style={{ color: "#b11b1b" }} variant="h5" align="center">Poseidon OS is not Mounted !!!</Typography>
-                        ) : null}
-                      </Grid>
+                      <ArrayCreate
+                        ssds={this.props.ssds}
+                        diskDetails={this.props.diskDetails}
+                        getDiskDetails={this.props.Get_Disk_Details}
+                      />
                     </Grid>
                   </Route>
                   <Route path="/storage/array/manage*">
                     <>
                       {this.props.arrayMap[this.props.selectedArray] ? (
-                        <React.Fragment>
-                          <Grid container spacing={1} className={classes.card}>
-                            <Grid item xs={12}>
-                              <Paper spacing={3} className={classes.spaced}>
-                                <Grid container className={classes.arrayInfoContainer} justifyContent="space-between">
-                                  <Tooltip
-                                    title={(
-                                      <Typography data-testid="array-id-text">
-                                        {`Array ID: ${this.props.arrayMap[this.props.selectedArray].uniqueId}`}
-                                      </Typography>
-                                    )}
-                                    classes={{
-                                      tooltip: classes.tooltip,
-                                    }}
-                                    interactive
-                                  >
-                                    <InfoIcon
-                                      data-testid="array-info-icon"
-                                      color="primary"
-                                      className={classes.arrayInfoIcon}
-                                    />
-                                  </Tooltip>
-                                  <Grid container className={classes.arraySelectStatus}>
-                                    <Grid item xs={12} sm={6} lg={4} className={classes.arraySelectGrid}>
-                                      <FormControl
-                                        className={classes.selectForm}
-                                        data-testid="arrayshow-form"
-                                      >
-                                        <InputLabel htmlFor="select-array">Select Array</InputLabel>
-                                        <Select
-                                          inputProps={{
-                                            id: "select-array",
-                                            "data-testid": "select-array-input",
-                                          }}
-                                          SelectDisplayProps={{
-                                            style: {
-                                              overflow: "hidden",
-                                              textOverflow: "ellipsis"
-                                            },
-                                            "data-testid": "select-array",
-                                          }}
-                                          onChange={this.changeArray}
-                                          value={this.props.selectedArray}
-                                          className={classes.arraySelect}
-                                        >
-                                          {this.props.arrays.map((array) => (
-                                            <MenuItem key={array.arrayname} value={array.arrayname}>
-                                              <Typography color="secondary">{array.arrayname}</Typography>
-                                            </MenuItem>
-                                          ))}
-                                        </Select>
-                                      </FormControl>
-                                    </Grid>
-                                    <Grid item xs={12} sm={6} lg={4} className={classes.arraySelectGrid}>
-                                      <Typography className={classes.statusText}>Status:
-                                        <span
-                                          style={{
-                                            fontWeight: 600,
-                                            color: this.props.arrayMap[this.props.selectedArray].status === "Mounted" ? "green" : "orange"
-                                          }}
-                                        >
-                                          {this.props.arrayMap[this.props.selectedArray].status}
-                                        </span>
-                                        <Grid container alignItems="center">, <span data-testid="array-show-status">{this.props.arrayMap[this.props.selectedArray].situation}</span>
-                                          {this.props.arrayMap[this.props.selectedArray].situation === "REBUILDING" ? (
-                                            <LightTooltip
-                                              data-testid="Tooltip"
-                                              TransitionComponent={Zoom}
-                                              title={(
-                                                <RebuildProgress
-                                                  arrayMap={this.props.arrayMap}
-                                                  array={this.props.selectedArray}
-                                                  progress={this.props.arrayMap[this.props.selectedArray].rebuildProgress}
-                                                  rebuildTime={this.props.arrayMap[this.props.selectedArray].rebuildTime}
-                                                />
-                                              )}
-                                              interactive
-                                            >
-                                              <InfoIcon
-                                                color="primary"
-                                                data-testid="rebuild-popover-icon"
-                                              />
-                                            </LightTooltip>
-                                          ) : null
-                                          }
-                                          {this.props.arrayMap[this.props.selectedArray].status === "Mounted" &&
-                                            this.props.arrayMap[this.props.selectedArray].situation === "DEGRADED" ? (
-                                            <Tooltip
-                                              title="Rebuild Array"
-                                            >
-                                              <Button
-                                                color="primary"
-                                                variant="contained"
-                                                data-testid="rebuild-icon"
-                                                id="rebuild-icon"
-                                                className={classes.rebuildButton}
-                                                onClick={this.rebuildArray}
-                                              >
-                                                Rebuild
-                                              </Button>
-                                            </Tooltip>
-                                          ) : null}
-                                        </Grid>
-                                      </Typography>
-                                    </Grid>
-                                  </Grid>
-                                  <ArrayShow
-                                    RAIDLevel={this.props.arrayMap[this.props.selectedArray].RAIDLevel}
-                                    slots={this.props.ssds}
-                                    arrayName={this.props.selectedArray}
-                                    corrupted={this.props.arrayMap[this.props.selectedArray].corrupted}
-                                    storagedisks={this.props.arrayMap[this.props.selectedArray].storagedisks}
-                                    sparedisks={this.props.arrayMap[this.props.selectedArray].sparedisks}
-                                    metadiskpath={this.props.arrayMap[this.props.selectedArray].metadiskpath}
-                                    writebufferdisks={this.props.arrayMap[this.props.selectedArray].writebufferdisks}
-                                    deleteArray={this.deleteArray}
-                                    writeThrough={this.props.arrayMap[this.props.selectedArray].writeThroughEnabled}
-                                    diskDetails={this.props.diskDetails}
-                                    getDiskDetails={this.props.Get_Disk_Details}
-                                    isDevicesFetching={this.props.isDevicesFetching}
-                                    isArrayInfoFetching={this.props.isArrayInfoFetching}
-                                    // detachDisk={this.props.Detach_Disk}
-                                    // attachDisk={this.props.Attach_Disk}
-                                    addSpareDisk={this.props.Add_Spare_Disk}
-                                    removeSpareDisk={this.props.Remove_Spare_Disk}
-                                    replaceDevice={this.props.Replace_Device}
-                                    mountStatus={this.props.arrayMap[this.props.selectedArray].status}
-                                    handleUnmountPOS={this.props.Unmount_POS}
-                                    handleMountPOS={this.props.Mount_POS}
-                                    getArrayInfo={this.props.Get_Array_Info}
-                                    getDevices={this.props.Get_Devices}
-                                  />
-                                </Grid>
-                              </Paper>
-                            </Grid>
-                          </Grid>
-                          <Grid
-                            container
-                            spacing={1}
-                            className={classes.card}
-                            style={{
-                              opacity: this.props.arrayMap[this.props.selectedArray].status !== "Mounted" ? 0.5 : 1,
-                              pointerEvents:
-                                this.props.arrayMap[this.props.selectedArray].status !== "Mounted"
-                                  ? "none"
-                                  : "initial",
-                            }}
-                          >
-                            <Grid item xs={12} md={6} className={classes.spaced}>
-                              <CreateVolume
-                                data-testid="createvolume"
-                                createVolume={this.createVolume}
-                                subsystems={this.props.subsystems}
-                                array={this.props.selectedArray}
-                                maxVolumeCount={this.props.maxVolumeCount}
-                                volCount={this.props.volumes.length}
-                                maxAvailableSize={
-                                  this.props.arrayMap[this.props.selectedArray].totalsize - totalVolSize
-                                }
-                                createVolSocket={this.state.createVolSocket}
-                                fetchVolumes={this.fetchVolumes}
-                                fetchArray={this.props.Get_Array}
-                                fetchSubsystems={this.props.Get_Subsystems}
-                              />
-                            </Grid>
-
-                            <Grid item xs={12} md={6}>
-                              <Paper className={classes.volumeStatsPaper}>
-                                <Typography className={classes.cardHeader}>
-                                  Volume Statistics
-                                </Typography>
-                                <div className={classes.statsWrapper}>
-                                  <Grid item xs={12}>
-                                    <Typography color="secondary">
-                                      Number of volumes: {this.props.volumes.length}
-                                    </Typography>
-                                  </Grid>
-                                  <Grid item xs={12} className={classes.statsContainer}>
-                                    <Box className={classes.volumeStats}>
-                                      <div style={volumeFilledStyle} />
-                                      <div style={volumeFreeStyle} />
-                                    </Box>
-                                    <Grid
-                                      item
-                                      container
-                                      xs={12}
-                                      wrap="wrap"
-                                      className={classes.legendContainer}
-                                    >
-                                      <Legend
-                                        bgColor="rgba(51, 158, 255,0.6)"
-                                        title={`
-                          Used Space :
-                          ${formatBytes(totalVolSize)}
-                        `}
-                                      />
-                                      <Legend
-                                        bgColor="rgba(0, 186, 0, 0.6)"
-                                        title={`
-                          Available for Volume Creation :
-                          ${formatBytes(
-                                          this.props.arrayMap[this.props.selectedArray].totalsize - totalVolSize >= BYTE_FACTOR * BYTE_FACTOR ?
-                                            this.props.arrayMap[this.props.selectedArray].totalsize - totalVolSize :
-                                            0
-                                        )}
-                        `}
-                                      />
-                                    </Grid>
-                                  </Grid>
-                                </div>
-                              </Paper>
-                            </Grid>
-                          </Grid>
-
-                          <Grid
-                            container
-                            spacing={1}
-                            className={classes.card}
-                            style={{
-                              opacity: this.props.arrayMap[this.props.selectedArray].status !== "Mounted" ? 0.5 : 1,
-                              pointerEvents:
-                                this.props.arrayMap[this.props.selectedArray].status !== "Mounted"
-                                  ? "none"
-                                  : "initial",
-                            }}
-                          >
-                            <Grid item xs={12}>
-                              <VolumeList
-                                ref={this.child}
-                                volumeFetch={this.fetchVolumes}
-                                volumes={this.props.volumes}
-                                fetchingVolumes={this.props.fetchingVolumes}
-                                deleteVolumes={this.deleteVolumes}
-                                resetQoS={this.props.Reset_Volume_QoS}
-                                editVolume={this.props.Edit_Volume}
-                                changeField={this.props.Change_Volume_Field}
-                                fetchVolumes={this.fetchVolumes}
-                                saveVolume={this.props.Reset_And_Update_Volume}
-                                changeMountStatus={this.changeMountStatus}
-                                changeMinType={this.props.Change_Min_Type}
-                                changeResetType={this.props.Change_Reset_Type}
-                              />
-                            </Grid>
-                          </Grid>
-                        </React.Fragment>
+                        <ArrayManage
+                          changeArray={this.changeArray}
+                          selectedArray={this.props.selectedArray}
+                          ssds={this.props.ssds}
+                          diskDetails={this.props.diskDetails}
+                          getDiskDetails={this.props.Get_Disk_Details}
+                          getDevices={this.props.Get_Devices}
+                          mountConfirm={this.mountConfirm}
+                        />
                       ) : null}
                     </>
                   </Route>
@@ -775,14 +307,7 @@ class StorageManagement extends Component {
 const mapStateToProps = (state) => {
   return {
     ssds: state.storageReducer.ssds,
-    metadisks: state.storageReducer.metadisks,
-    volumes: state.storageReducer.volumes,
-    fetchingVolumes: state.storageReducer.fetchingVolumes,
-    isDevicesFetching: state.storageReducer.isDevicesFetching,
-    isArrayInfoFetching: state.storageReducer.isArrayInfoFetching,
-    arrays: state.storageReducer.arrays,
     arrayMap: state.storageReducer.arrayMap,
-    config: state.storageReducer.config,
     selectedArray: state.storageReducer.arrayname,
     loading: state.storageReducer.loading,
     subsystems: state.subsystemReducer.subsystems,
@@ -794,68 +319,33 @@ const mapStateToProps = (state) => {
     errorMsg: state.storageReducer.errorMsg,
     errorCode: state.storageReducer.errorCode,
     arraySize: state.storageReducer.arraySize,
-    maxVolumeCount: state.storageReducer.maxVolumeCount,
     totalVolSize: state.storageReducer.totalVolSize,
     slots: state.storageReducer.slots,
     arrayExists: state.storageReducer.arrayExists,
-    selectedRaid: state.storageReducer.selectedRaid,
     RAIDLevel: state.storageReducer.RAIDLevel,
     diskDetails: state.storageReducer.diskDetails,
     loadText: state.storageReducer.loadText,
     mountStatus: state.storageReducer.mountStatus,
-    posMountStatus: state.headerReducer.state
   };
 };
 const mapDispatchToProps = (dispatch) => {
   return {
     Get_Devices: (payload) =>
       dispatch({ type: actionTypes.SAGA_FETCH_DEVICE_INFO, payload }),
-    Create_Volume: (payload) =>
-      dispatch({ type: actionTypes.SAGA_SAVE_VOLUME, payload }),
-    Get_Array: () => dispatch({ type: actionTypes.SAGA_FETCH_ARRAY }),
-    Delete_Array: (payload) =>
-      dispatch({ type: actionTypes.SAGA_DELETE_ARRAY, payload }),
     Get_Volumes: (payload) => dispatch({ type: actionTypes.SAGA_FETCH_VOLUMES, payload }),
-    Get_Config: () => dispatch({ type: actionTypes.SAGA_FETCH_CONFIG }),
-    Delete_Volumes: (payload) =>
-      dispatch({ type: actionTypes.SAGA_DELETE_VOLUMES, payload }),
+
     Close_Alert: () => dispatch({ type: actionTypes.STORAGE_CLOSE_ALERT }),
-    Create_Array: (payload) =>
-      dispatch({ type: actionTypes.SAGA_CREATE_ARRAY, payload }),
     Get_Disk_Details: (payload) =>
       dispatch({ type: actionTypes.SAGA_FETCH_DEVICE_DETAILS, payload }),
-    Edit_Volume: (payload) =>
-      dispatch({ type: actionTypes.EDIT_VOLUME, payload }),
     Change_Mount_Status: (payload) =>
       dispatch({ type: actionTypes.SAGA_VOLUME_MOUNT_CHANGE, payload }),
-    Change_Volume_Field: (payload) =>
-      dispatch({ type: actionTypes.CHANGE_VOLUME_FIELD, payload }),
-    Change_Min_Type: (payload) =>
-      dispatch({ type: actionTypes.CHANGE_MIN_TYPE, payload }),
-    Change_Reset_Type: (payload) =>
-      dispatch({ type: actionTypes.CHANGE_RESET_TYPE, payload }),
-    Reset_And_Update_Volume: (payload) =>
-      dispatch({ type: actionTypes.SAGA_RESET_AND_UPDATE_VOLUME, payload }),
     // Detach_Disk: (payload) => dispatch({ type: actionTypes.SAGA_DETACH_DISK, payload }),
     // Attach_Disk: (payload) => dispatch({ type: actionTypes.SAGA_ATTACH_DISK, payload }),
-    Add_Spare_Disk: (payload) =>
-      dispatch({ type: actionTypes.SAGA_ADD_SPARE_DISK, payload }),
-    Remove_Spare_Disk: (payload) =>
-      dispatch({ type: actionTypes.SAGA_REMOVE_SPARE_DISK, payload }),
-    Replace_Device: (payload) =>
-      dispatch({ type: actionTypes.SAGA_REPLACE_DEVICE, payload }),
     Get_Max_Volume_Count: () =>
       dispatch({ type: actionTypes.SAGA_FETCH_MAX_VOLUME_COUNT }),
-    Unmount_POS: () => dispatch({ type: actionTypes.SAGA_UNMOUNT_POS }),
-    Mount_POS: (payload) => dispatch({ type: actionTypes.SAGA_MOUNT_POS, payload }),
     Set_Array: (payload) => dispatch({ type: actionTypes.SET_ARRAY, payload }),
-    Select_Raid: (payload) => dispatch({ type: actionTypes.SELECT_RAID, payload }),
-    Get_Subsystems: () => dispatch({ type: actionTypes.SAGA_FETCH_SUBSYSTEMS }),
     Show_Storage_Alert: (payload) => dispatch({ type: actionTypes.STORAGE_SHOW_ALERT, payload }),
-    Auto_Create_Array: (payload) => dispatch({ type: actionTypes.SAGA_AUTO_CREATE_ARRAY, payload }),
-    Reset_Volume_QoS: (payload) => dispatch({ type: actionTypes.SAGA_RESET_VOLUME_QOS, payload }),
-    Get_Array_Info: (payload) => dispatch({ type: actionTypes.SAGA_GET_ARRAY_INFO, payload }),
-    Rebuild_Array: (payload) => dispatch({ type: actionTypes.SAGA_REBUILD_ARRAY, payload })
+    Get_Subsystems: () => dispatch({ type: actionTypes.SAGA_FETCH_SUBSYSTEMS }),
   };
 };
